requirements.txt
tensorflow
numpy
pillow
fastapi
uvicorn